requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.13
django-admin-list-filter-dropdown==1.0.3
django-json-widget==2.0.1
djangorestframework==3.15.2
exceptiongroup==1.2.1
flake8==7.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
kombu==5.3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
telegram-django-bot==1.1.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
